Move TimeSpanGroupBuilder and assert_response_status_code under tests/utils.py
